# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": " NodeJS API project to demonstrate Swagger ",
        "version": "1.0.0",
        "title": "E-commers Project API"
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:7065",
    "basePath": "/api/v1",
    "paths": {
...
        "/add-product/product": {
            "post": {
                "summary": "Save the Product",
                "description": "Create the Product ",
                "produces": [
                    "application/json"
                ],
                "requestBody": true,
                "parameters": [{
                        "in": "body",
                        "name": "body",
                        "description": "Product object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "product_Name": {
                                    "type": "string",
                                    "example": "mobile"
                                },
                                "description": {
                                    "type": "String",
                                    "example": "Good"
                                },

                                "category": {
                                    "type": "String",
                                    "example": "tab"
                                },

                                "price": {
                                    "type": "Number",
                                    "example": "7600"
                                },

                                "stock": {
                                    "type": "Number",
                                    "example": "80"
                                },
                            
                                "image_URL": {
                                    "type": "String",
                                    "example": "mobile"
                                }
                            }
                        }
                },
                    
                {
                    "name": "Authorization",
                    "in": "header",
                    "description": "token",
                    "required": true,
                    "type": "string"
                }
            ],
                "responses": {
                    "201": {
                        "description": "successful operation ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/productResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/product/get-all-product": {
            "get": {
                "summary": "Get All  Product",
                "description": "Get All  Product !!!",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    
                    {
                        "name": "product_Name",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "stock",
                        "in": "query",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "image_URL",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/productResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/product/get-product/{product_Id}": {
            "get": {
                "summary": "Get One Session",
                "description": "Get One Session",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "product_Id",
                        "in": "path",
                        "description": "product_Id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/productResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/product/update-product/{product_Id}": {
            "put": {
                "summary": "Update the Product",
                "description": "Update the Product ",
                "produces": [
                    "application/json"
                ], 
                "consumes": ["application/json"],
                "requestBody": true,
                "parameters": [
                    {
                        "in": "path",
                        "name": "product_Id",
                        "description": "Product_Id",
                        "required": true,
                        "type": "string"
                     
                    },
                    
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "updateData",
                        "in": "body",
                        "description": " body ",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "product_Name": {
                                    "type": "string",
                                    "example": "mobile"
                                },
                                "description": {
                                    "type": "String",
                                    "example": "Good"
                                },

                                "category": {
                                    "type": "String",
                                    "example": "tab"
                                },

                                "price": {
                                    "type": "Number",
                                    "example": "7600"
                                },

                                "stock": {
                                    "type": "Number",
                                    "example": "80"
                                },
                            
                                "image_URL": {
                                    "type": "String",
                                    "example": "mobile"
                                }
                            }                        
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/productResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/product/delete-product/{product_Id}": {
            "delete": {
                "summary": "Delete the product",
                "description": "Delete the product",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "product_Id",
                        "in": "path",
                        "description": "product_Id",
                        "required": true,
                        "type": "string"
                    },
                    
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/deleteResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "productResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "true"
                },
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "data": {
                    "type": "object",
                   
                        "properties": {
                            "product_Name": {
                                "type": "string",
                                "example": "mobile"
                            },
                            "description": {
                                "type": "String",
                                "example": "very good phone"
                            },

                            "category": {
                                "type": "String",
                                "example": "tab"
                            },

                            "price": {
                                "type": "Number",
                                "example": "5400"
                            },

                            "stock": {
                                "type": "Number",
                                "example": "57"
                            },
                        
                            "image_URL": {
                                "type": "String",
                                "example": "mobile"
                            },
                            "createdAt": { "type": "string", "format": "date-time", "example": "2023-11-30T07:52:22.910Z" },
                            "updatedAt": { "type": "string", "format": "date-time", "example": "2023-11-30T07:52:22.910Z" }
                        }
                
                    
                }
            }
        },
...
        "deleteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Deleted successfully"
                },
                "Session": {
                    "type": "object",
                    "properties": {
                        "acknowledged": {
                            "type": "string",
                            "example": "true"
                        },
                        "deletedCount": {
                            "type": "number",
                            "example": "2"
                        }
                    }
                }
            }
        },
...
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "Number",
                    "example": "400"
                   
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}